pyproject.toml
[tool.poetry]
name = "home-automation"
version = "0.1.0"
description = ""
authors = ["Matthew Gilpin <matt@matthewgilpin.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
influxdb-client = "^1.44.0"
python-dotenv = "^1.0.1"
pywizlight = "^0.5.14"
paho-mqtt = "^2.1.0"
pexpect = "^4.9.0"
aranet4 = "^2.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
